We can push the DotWrapper down into Dot
