+++ common/views/components/Dot/Dot.tsx
@@ -1,6 +1,13 @@
 import styled from 'styled-components';
 import { FunctionComponent } from 'react';
 import { PaletteColor } from '@weco/common/views/themes/config';
+import Space from '../styled/Space';
+
+const DotWrapper = styled(Space).attrs({
+  as: 'span',
+  h: { size: 'xs', properties: ['margin-right'] },
+  className: 'flex flex--v-center',
+})``;
 
 const DotEl = styled.span.attrs({
   'aria-hidden': true,
@@ -18,6 +25,8 @@ type Props = {
 };
 
 const Dot: FunctionComponent<Props> = ({ dotColor }) => (
-  <DotEl dotColor={dotColor} />
+  <DotWrapper>
+    <DotEl dotColor={dotColor} />
+  </DotWrapper>
 );
 export default Dot;
