+++ content/webapp/components/StatusIndicator/StatusIndicator.tsx
@@ -10,7 +10,6 @@ import {
   today,
 } from '@weco/common/utils/dates';
 import { PaletteColor } from '@weco/common/views/themes/config';
-import DotWrapper from '@weco/common/views/components/Dot/DotWrapper';
 
 type Props = {
   start: Date;
@@ -52,9 +51,7 @@ const StatusIndicator: FunctionComponent<Props> = ({
 
   return (
     <span className={`flex flex--v-center ${font('intr', 5)}`}>
-      <DotWrapper>
-        <Dot dotColor={color} />
-      </DotWrapper>
+      <Dot dotColor={color} />
       <span>{text}</span>
     </span>
   );
